arithub triage: fix 4 real signed-overflow bugs, confirm rest is proof-power gap
Manually sampled ~45 of the ~430 unproved SignedArithmetic findings
lint.sh arithub still reports on x86_64/aarch64/i386, across 14 files,
weighted away from the files the prior investigation already traced
(sed.c, awk_run.c). Most are confirmed proof-power gap: the guard that
makes the arithmetic safe exists, just not in a form the checker's Z3
modeling can correlate with the flagged expression (a prior statement's
invariant on a global, an interprocedural precondition established by
a caller, a nonlinear division-based overflow guard next to the
multiply it protects).

Four were real, confirmed undefined behaviour on directly reachable
user input, not proof-power gap, and are fixed here:

- stdlib/abs.c: abs()/labs()/llabs()/imaxabs() negated their argument
  with a plain unary '-', undefined at INT_MIN (etc.). Now negate via
  unsigned wraparound and cast back, matching stdio_impl.h's existing
  __neg_mag() idiom for the identical problem.

- util/expr.c: do_arith() ran +, -, *, /, % on two operands taken
  straight from argv (e.g. "expr 9223372036854775807 + 1") with no
  overflow guard at all. Added CERT-style pre-checks that report
  "expr: overflow" through the file's existing xerr() diagnostic path,
  the same way "division by zero" was already handled.

- util/patch.c: parse_uint(), which reads a hunk header's line
  numbers/counts straight out of whatever patch file the process was
  handed, accumulated digits with an unguarded v = v*10+digit. Now
  rejects an overflowing header the same way it already rejects a
  non-digit one.

- util/man.c: the .if conditional-expression evaluator and the
  .nr/\n+x register auto-increment arithmetic -- both processing
  numbers embedded in man page source, a realistic untrusted-input
  surface -- ran unary '-', +, -, *, /, % on `long` with no guard.
  Added small wrap-arithmetic helpers (man_wrap_add/sub/mul/neg)
  following the same unsigned-wraparound idiom as abs.c, preserving
  troff's existing "silently produce a number" behaviour instead of
  introducing a new error path this evaluator did not have.

Verified: all four files build clean (aarch64-linux, clang); the
multiplication/addition/subtraction overflow-check logic itself was
exhaustively checked against __int128 arithmetic (0 mismatches over
~4M cases). Re-running lint.sh arithub confirms every finding at the
fixed sites is gone; the do_arith() multiplication and parse_uint's
multiply-add remain flagged even after the fix, for the same
proof-power reason printf.c's pre-existing, structurally identical
digit-accumulation guard is also flagged -- not a sign the fix is
wrong, and not something to chase by touching the checker itself.

+++ src/stdlib/abs.c
@@ -8,9 +8,15 @@
 #include <stdlib.h>
 #include <inttypes.h>
 
-int abs(int a) { return a > 0 ? a : -a; }
-long labs(long a) { return a > 0 ? a : -a; }
-long long llabs(long long a) { return a > 0 ? a : -a; }
-intmax_t imaxabs(intmax_t a) { return a > 0 ? a : -a; }
+/* Negating a might overflow: INT_MIN (etc.) has no representable
+ * positive counterpart, so a plain `-a` is undefined right at the one
+ * input where callers most need a defined answer. Negating the
+ * unsigned reinterpretation instead wraps modulo 2**N (C99 6.2.5p9),
+ * giving the same bit pattern every existing caller already relies on
+ * (abs(INT_MIN) == INT_MIN) without the intermediate UB. */
+int abs(int a) { return a > 0 ? a : (int)-(unsigned)a; }
+long labs(long a) { return a > 0 ? a : (long)-(unsigned long)a; }
+long long llabs(long long a) { return a > 0 ? a : (long long)-(unsigned long long)a; }
+intmax_t imaxabs(intmax_t a) { return a > 0 ? a : (intmax_t)-(uintmax_t)a; }
 
 // NOLINTEND(misc-include-cleaner)
